+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFJQ5Y5ZCMJYMF52BT7J8A4C"
+        "01KFJYQGREH929J0ZBBR5EDQGX"
     )
     r"""The unique identifier of the data source"""