+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFJQ5Y5VPWTG2ZWAS4Q83YW3"
+        "01KFJYQGRA1W6A4ZSVM42XVSJP"
     )
     r"""The unique identifier of the data source"""