+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFJQ5Y5YRESXSXM7PRHR6QA9"
+        "01KFJYQGRC7R19CC12YK27BCZN"
     )
     r"""The unique identifier of the data source"""