+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFJQ5Y63KHYZDZVJ4PCXMRSV"
+        "01KFJYQGRJXPFXY227CQJ3F8QR"
     )
     r"""The unique identifier of the data source"""